cam_http_test/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
aiofiles==23.1.0
matplotlib==3.8.0
pillow==10.0.0
python-multipart==0.0.6